refactor: move memberTypes module to the own folder
